separate class and function into dedicated files. Changed scraper to iterate over full list of keywords and output a dictionary. Implemented json file for keywords
