+++ blueskyScraper.py
@@ -1,34 +1,9 @@
-from playwright.sync_api import sync_playwright, Playwright
+from playwright.sync_api import Playwright
 import time
 from nltk.sentiment import SentimentIntensityAnalyzer
 import nltk
 import re
-
-class SkeetData:
-    link_headline = None
-    link_url = None
-    quoted_user = None
-    quoted_text = None
-
-    def __init__(
-            self,
-            post_date,
-            username,
-            handle,
-            text_content,
-            comments,
-            reposts,
-            likes,
-            sentiment
-    ):
-        self.post_date = post_date
-        self.username = username
-        self.handle = handle
-        self.text_content = text_content
-        self.comments = comments
-        self.reposts = reposts
-        self.likes = likes
-        self.sentiment = sentiment
+from postClasses import SkeetData
 
 SKEET_FINDER_PATH = "html > body > div:nth-child(1) > div > div > div > div > div > main > div > div > div:nth-child(2) > div > div > div:nth-child(2) > div:nth-child(3) > div > div:nth-child(2) > div > div > div:nth-child(2) > div > div:nth-child"
 SKEET_CONTENT_PATH = "div > div:nth-child(2) > div:nth-child(2)"
@@ -67,10 +42,13 @@ def DateParser(date: str):
         if amPmSplitter[1] == "PM":
             if int(hourMinuteSplit[0]) < 12:
                 hourMinuteSplit[0] = str(int(hourMinuteSplit[0])+12)
+    # TODO: convert to UTC
     return dateSplitter[0]+ " " + hourMinuteSplit[0] + ":" + hourMinuteSplit[1]
 
 
-def run(playwright: Playwright, searchQuery: str):
+def blueskyScraper(playwright: Playwright, searchQueries, postDepth: int):
+    allQueriesHolder = {}
+    postHolder = list()
     chromium = playwright.chromium
     browser = chromium.launch(headless=False)
     page = browser.new_page()
@@ -83,67 +61,69 @@ def run(playwright: Playwright, searchQuery: str):
     page.locator("button[aria-label='Next']").click()
     page.wait_for_load_state("networkidle")
     searchBox = page.get_by_role("search")
-    searchBox.fill(searchQuery)
-    searchBox.press("Enter")
-    page.wait_for_load_state("networkidle")
-    for skeets in range(2,3):
-        linkObject = {"headline": "", "link": ""}
-        quoteSkeet = {"quotedUser": "", "quotedText": ""}
-        finderPath = SKEET_FINDER_PATH + "(" + str(skeets) + ")"
-        contentPath = finderPath + " > " + SKEET_CONTENT_PATH
-        handlePath = contentPath + " > " + SKEET_USER_HANDLE_PATH
-        handleText = page.locator(handlePath).text_content()
-        username = handleText.split("@")[0]
-        usernameStepOne = re.sub(r"[\xa0]", "", username)
-        usernameClean = re.sub(r"[\u202a-\u202e]", "", usernameStepOne)
-        handle = handleText.split("@")[1]
-        handleClean = "@" + re.sub(r"[\u202a-\u202e]", "", handle)
-        timestampPath = contentPath + " > " + SKEET_TIMESTAMP_PATH
-        timestamp = page.locator(timestampPath).get_attribute("aria-label")
-        timestampClean = DateParser(timestamp)
-        if (page.locator(contentPath + " > div:nth-child(4)").is_visible()):
-            textBuff = " > div:nth-child(3)"
-            interactionBuff = " > div:nth-child(4) > "
-            textPath = contentPath + textBuff
-            commentsPath = contentPath + interactionBuff + SKEET_COMMENTS_PATH
-            reskeetsPath = contentPath + interactionBuff + SKEET_RESKEETS_PATH
-            likesPath = contentPath + interactionBuff + SKEET_LIKES_PATH
-        else:
-            textPath = contentPath + " > " + SKEET_TEXT_PATH
-            commentsPath = contentPath + " > div:nth-child(3) > " + SKEET_COMMENTS_PATH
-            reskeetsPath = contentPath + " > div:nth-child(3) > " + SKEET_RESKEETS_PATH
-            likesPath = contentPath + " > div:nth-child(3) > " + SKEET_LIKES_PATH
-        if (page.locator(textPath + " > div:nth-child(2)").is_visible()):
-            if (page.locator(textPath + " > div:nth-child(2) > a").is_visible()):
-                linkObject["headline"] = page.locator(textPath + " > div:nth-child(2) > a").get_attribute("aria-label")
-                linkObject["link"] = page.locator(textPath + " > div:nth-child(2) > a").get_attribute("href")
-            elif (page.locator(textPath + " > div:nth-child(2) > div > div:nth-child(2)[role='link']").is_visible()):
-                quoteUserWithFormatting = page.locator(textPath + " > " + QUOTE_SKEET_HANDLE_PATH + " > a:nth-child(1)").text_content()
-                quoteUserClean = re.sub(r"[\u202a-\u202e]", "", quoteUserWithFormatting)
-                quoteSkeet["quotedUser"] = quoteUserClean
-                quoteSkeet["quotedText"] = page.locator(textPath + " > " + QUOTE_SKEET_TEXT_PATH).text_content()
-            textPath = textPath + " > div:nth-child(1)"
-        skeetText = page.locator(textPath).text_content()
-        skeetTextClean = re.sub(r"(\n+)", " ", skeetText)
-        sentimentScores = sia.polarity_scores(skeetText)
-        commentsCount = page.locator(commentsPath).text_content()
-        commentsInt = CountConverter(commentsCount)
-        reskeetsCount = page.locator(reskeetsPath).text_content()
-        reskeetsInt = CountConverter(reskeetsCount)
-        likesCount = page.locator(likesPath).text_content()
-        likesInt = CountConverter(likesCount)
-        sendToDB = SkeetData(timestampClean, usernameClean, handleClean, skeetTextClean, commentsInt, reskeetsInt, likesInt, sentimentScores["compound"])
-        if (list(linkObject.values())[0]):
-            sendToDB.link_headline = linkObject["headline"]
-            sendToDB.link_url = linkObject["link"]
-        if (list(quoteSkeet.values())[0]):
-            sendToDB.quoted_user = quoteSkeet["quotedUser"]
-            sendToDB.quoted_text = quoteSkeet["quotedText"]
-        if (skeets%20 == 0):
-            page.evaluate("()=> window.scrollBy(0, 10000)")
-            page.wait_for_timeout(500)
+    for searchQuery in searchQueries:
+        searchBox.fill(searchQuery)
+        searchBox.press("Enter")
+        page.wait_for_load_state("networkidle")
+        for skeets in range(2, postDepth+2):
+            linkObject = {"headline": "", "link": ""}
+            quoteSkeet = {"quotedUser": "", "quotedText": ""}
+            finderPath = SKEET_FINDER_PATH + "(" + str(skeets) + ")"
+            contentPath = finderPath + " > " + SKEET_CONTENT_PATH
+            handlePath = contentPath + " > " + SKEET_USER_HANDLE_PATH
+            handleText = page.locator(handlePath).text_content()
+            username = handleText.split("@")[0]
+            usernameStepOne = re.sub(r"[\xa0]", "", username)
+            usernameClean = re.sub(r"[\u202a-\u202e]", "", usernameStepOne)
+            handle = handleText.split("@")[1]
+            handleClean = "@" + re.sub(r"[\u202a-\u202e]", "", handle)
+            timestampPath = contentPath + " > " + SKEET_TIMESTAMP_PATH
+            timestamp = page.locator(timestampPath).get_attribute("aria-label")
+            timestampClean = DateParser(timestamp)
+            if (page.locator(contentPath + " > div:nth-child(4)").is_visible()):
+                textBuff = " > div:nth-child(3)"
+                interactionBuff = " > div:nth-child(4) > "
+                textPath = contentPath + textBuff
+                commentsPath = contentPath + interactionBuff + SKEET_COMMENTS_PATH
+                reskeetsPath = contentPath + interactionBuff + SKEET_RESKEETS_PATH
+                likesPath = contentPath + interactionBuff + SKEET_LIKES_PATH
+            else:
+                textPath = contentPath + " > " + SKEET_TEXT_PATH
+                commentsPath = contentPath + " > div:nth-child(3) > " + SKEET_COMMENTS_PATH
+                reskeetsPath = contentPath + " > div:nth-child(3) > " + SKEET_RESKEETS_PATH
+                likesPath = contentPath + " > div:nth-child(3) > " + SKEET_LIKES_PATH
+            if (page.locator(textPath + " > div:nth-child(2)").is_visible()):
+                if (page.locator(textPath + " > div:nth-child(2) > a").is_visible()):
+                    linkObject["headline"] = page.locator(textPath + " > div:nth-child(2) > a").get_attribute("aria-label")
+                    linkObject["link"] = page.locator(textPath + " > div:nth-child(2) > a").get_attribute("href")
+                elif (page.locator(textPath + " > div:nth-child(2) > div > div:nth-child(2)[role='link']").is_visible()):
+                    quoteUserWithFormatting = page.locator(textPath + " > " + QUOTE_SKEET_HANDLE_PATH + " > a:nth-child(1)").text_content()
+                    quoteUserClean = re.sub(r"[\u202a-\u202e]", "", quoteUserWithFormatting)
+                    quoteSkeet["quotedUser"] = quoteUserClean
+                    quoteSkeet["quotedText"] = page.locator(textPath + " > " + QUOTE_SKEET_TEXT_PATH).text_content()
+                textPath = textPath + " > div:nth-child(1)"
+            skeetText = page.locator(textPath).text_content()
+            skeetTextClean = re.sub(r"(\n+)", " ", skeetText)
+            sentimentScores = sia.polarity_scores(skeetText)
+            commentsCount = page.locator(commentsPath).text_content()
+            commentsInt = CountConverter(commentsCount)
+            reskeetsCount = page.locator(reskeetsPath).text_content()
+            reskeetsInt = CountConverter(reskeetsCount)
+            likesCount = page.locator(likesPath).text_content()
+            likesInt = CountConverter(likesCount)
+            sendToDB = SkeetData(timestampClean, usernameClean, handleClean, skeetTextClean, commentsInt, reskeetsInt, likesInt, sentimentScores["compound"])
+            if (list(linkObject.values())[0]):
+                sendToDB.link_headline = linkObject["headline"]
+                sendToDB.link_url = linkObject["link"]
+            if (list(quoteSkeet.values())[0]):
+                sendToDB.quoted_user = quoteSkeet["quotedUser"]
+                sendToDB.quoted_text = quoteSkeet["quotedText"]
+            if (skeets%20 == 0):
+                page.evaluate("()=> window.scrollBy(0, 10000)")
+                page.wait_for_timeout(500)
 
-        # TODO: setup pipeline to database
+            postHolder.append(sendToDB)
+        allQueriesHolder[searchQuery] = postHolder
+    return allQueriesHolder
 
-with sync_playwright() as playwright:
-    run(playwright, "police")
+
